Avoid importing client-zip in the native build.
